agent_api/requirements.txt
fastapi 
uvicorn 
pydantic 
google-generativeai 
fastmcp 
httpx
python-dotenv
google-adk